feat: implement agent stats retrieval in analytics module and update routes
